Add files via upload
Experience the thrill of drumming with the "Drum Kit Simulator." Click or use keyboard shortcuts to play seven different drum sounds. Whether you're a percussion pro or just looking for a rhythmic break, this project offers a quick and interactive way to create beats. Let the music flow with mouse and keyboard-friendly controls.

Technologies Used:

HTML: Structured the layout and elements.
CSS: Styled the user interface for an appealing design.
JavaScript: Implemented the drum sounds, mouse click, and keyboard interactions for a responsive and engaging drumming experience.